feat(MainInnerPortal): can be configured by passing additional props
isReady (Boolean) - tells InnerPortal if it should rely on external data to be fullfilled in global
state before rendering content or use settings from component it is wrapping. For example - we don't
need to wait until data will arrive and saved to store to display Dealer portal content
placeholders, so by passing `true` as the value we telling HoC to render component instantly.